+++ tht-apis/src/main/java/com/tht/thtapis/facade/user/UserFacade.java
@@ -1,5 +1,6 @@
 package com.tht.thtapis.facade.user;
 
+import com.tht.domain.entity.block.UserBlockService;
 import com.tht.enums.user.Gender;
 import com.tht.domain.entity.chat.service.ChatRoomUserService;
 import com.tht.domain.entity.dailyfalling.mapper.DailyFallingTimeMapper;
