+++ tht-core/domain/src/main/java/com/tht/domain/entity/block/dto/UserBlockDto.java
@@ -1,0 +1,4 @@
+package com.tht.domain.entity.block.dto;
+
+public class UserBlockDto {
+}
